requirements.txt
playwright
nest-asyncio
google
mysql-connector-python
python-dotenv
aiohttp
aiofiles
google-generativeai
openai
beautifulsoup4
Pillow
pytest